Merge in METIS/metis-fin-calc-service from release/MET-69-backend-add-swagger-documentation-and-api-release to master

* commit 'fc0f309eb2beb6ea64f1939722e3101e9b82cf8f':
  postgres reconfig
  InvestmentController documentation completed
  CategoryController documentation completed
  portfolioController documented
  templated all endpoints for swagger documentation, userController completed

+++ src/main/java/pl/fintech/metisfinancialcalculator/fincalcservice/service/UserService.java
@@ -52,6 +52,8 @@ public class UserService {
         }
     }
     public void delete(String username) {
+        if(null==userRepository.findByUsername(username))
+            throw new CustomException("The user doesn't exist", HttpStatus.NOT_FOUND);
         userRepository.deleteByUsername(username);
     }
     public User search(String username) {
